Create the program flow
How the program will flow for example, we start by setting the alarm and taking the input from the user then the clock counts till the time matches the the set alarm. After that, the LCD displays the riddle and waits for the input from the user. If it is correct, it displays correct otherwise displays wrong